backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
chromadb
python-dotenv
PyPDF2
tqdm
